requirements.txt
GitPython
boto3